fix saveToDouble

+++ util/shared/src/main/scala/StringOps.scala
@@ -16,7 +16,7 @@ object StringOps {
   @inline def safeToInt(intStr: String): Option[Int] = {
     try { Some(intStr.toInt) } catch { case NonFatal(_) => None }
   }
-  @inline def safeToDouble(doubleStr: String): Option[Int] = {
-    try { Some(doubleStr.toInt) } catch { case NonFatal(_) => None }
+  @inline def safeToDouble(doubleStr: String): Option[Double] = {
+    try { Some(doubleStr.toDouble) } catch { case NonFatal(_) => None }
   }
 }
